requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.0
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2
python-multipart==0.0.6
plivo==4.52.0